community/other/cloud-function-for-shared-vpcs/cloudfunction/requirements.txt
functions-framework==3.*
google-cloud-compute